+++ pyxel/models/charge_generation/__init__.py
@@ -13,5 +13,5 @@ from .dark_current_rule07 import dark_current_rule07
 from .load_charge import load_charge
 from .photoelectrons import simple_conversion, conversion_with_qe_map
 from .cosmix.cosmix import cosmix
-from .dark_current import dark_current, dark_current_saphira
+from .dark_current import dark_current, simple_dark_current, dark_current_saphira
 from .apd_gain import apd_gain